bot/requirements.txt
discord.py
tk